% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantF.R
\name{constant.f}
\alias{constant.f}
\title{constant.f}
\usage{
constant.f(m, t)
}
\arguments{
\item{m}{death rate of uninfected and pre-patent snails, a numeric number or vector}

\item{t}{pre-patent period, a numeric number or vector}
}
\value{
a numeric vector, the resulting constant for the main equation
}

