% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunction.R
\name{testfunction}
\alias{testfunction}
\title{testfunction}
\usage{
testfunction(X)
}
\arguments{
\item{X}{numeric value ##you need to have the variable first and then the description}
}
\value{
X*X
}
\description{
this function does X*X
}

